Refactor this out and add some tests
